server/requirements.txt
fastapi
uvicorn
httpx
python-multipart
pillow
python-dotenv
pydantic
firebase-admin
websockets
opencv-python
torch
diffusers
transformers
accelerate
controlnet-aux
mediapipe